server/requirements.txt
fastapi[all]
uvicorn
ultralytics
numpy
python-dotenv